prj_summer-badge/tool-py_send-qr/requirements.txt
mysql-connector-python==8.0.33
qrcode==7.4.2
